Revert "drivers: move eeprom_slave driver to tests directory"
This reverts commit cabbd916cfba20f39a91922b740d914f38a21b2d.

This is considered to be useful enough that it should be restored
as a stable Zephyr API.
