+++ doc/reference/peripherals/i2c_eeprom_slave.rst
@@ -1,0 +1,13 @@
+.. _i2c_eeprom_slave_api:
+
+I2C EEPROM Slave
+################
+
+Overview
+********
+
+API Reference
+**************
+
+.. doxygengroup:: i2c_eeprom_slave_api
+   :project: Zephyr
