+++ doc/reference/peripherals/index.rst
@@ -18,6 +18,7 @@ Peripherals
    gna.rst
    gpio.rst
    hwinfo.rst
+   i2c_eeprom_slave.rst
    i2c.rst
    ipm.rst
    kscan.rst
