+++ drivers/i2c/slave/CMakeLists.txt
@@ -1,1 +1,3 @@
 # SPDX-License-Identifier: Apache-2.0
+
+zephyr_sources_ifdef(CONFIG_I2C_EEPROM_SLAVE	eeprom_slave.c)
