+++ include/drivers/i2c/slave/eeprom.h
