+++ tests/drivers/i2c/i2c_slave_api/prj.conf
@@ -4,3 +4,5 @@ CONFIG_BOOT_BANNER=y
 CONFIG_ZTEST=y
 CONFIG_I2C=y
 CONFIG_I2C_SLAVE=y
+CONFIG_I2C_EEPROM_SLAVE=y
+CONFIG_I2C_VIRTUAL=y
